Combined RDT receiver and sender.
Sequence numbers are now synchronised between an RDT sender and RDT receiver.

This involved removing the RDTSender state machine (as it had trivially
only 2 states).

Added a new RDTConnection class that when instatiated creates a synchronised
RDTSender and RDTReceiver object that communicate via the connection object itself
(passed to them as their context)